Make mesh-heap bounds check 64-bit-safe (pointer hygiene, stage 3e)
CMesh::bInMeshHeap tests whether an arbitrary pointer lies inside the
global mesh fast heap by truncating pv and the heap base/end to uint32
and comparing. If the heap is mapped above 4GB, the truncated
absolute-address comparison gives the wrong answer. Widen to uintptr.

This is the same bug and fix as ImageLoader::bLoadHeapAllocation
(43c4b87); the two are twin heap-bounds predicates.

Mesh.cpp needs nothing: its one (int)uSize() cast is a byte count for a
memory-log counter, not a pointer.

Behaviour-identical on Win32 where uintptr == uint32: built green (exit 0,
no new warnings), smoke-launched OK. Fifth slice of the uint32->uintptr
sweep after VirtualMem (ba41823), FastHeap (3825cd3), loaders (43c4b87),
audio (1d560a9).

+++ jp2_pc/Source/Lib/GeomDBase/Mesh.hpp
@@ -168,8 +168,8 @@ public:
 	)
 	//*************************************
 	{
-		return ( ((uint32)pv <= ((uint32)fhGlobalMesh.uNumBytesUsed() + (uint32)fhGlobalMesh.pvGetBase())) &&
-		((uint32)pv >= (uint32)fhGlobalMesh.pvGetBase()) );
+		return ( ((uintptr)pv <= (fhGlobalMesh.uNumBytesUsed() + (uintptr)fhGlobalMesh.pvGetBase())) &&
+		((uintptr)pv >= (uintptr)fhGlobalMesh.pvGetBase()) );
 	}
